fix:   - The tcombLocalName used for guarding tcomb imports gets reset at the     start of each new File being processed.   - Additional logic added to `require` import resolution based on manual     testing on a large project. chore:   - Cleaned up the tests structure by merging test/guards.js into test/index.js.
